# Conflicts:
#	crates/bevy_app/src/app.rs
#	crates/bevy_app/src/app_builder.rs
#	crates/bevy_asset/src/lib.rs
#	crates/bevy_core/src/time/fixed_timestep.rs
#	crates/bevy_ecs/Cargo.toml
#	crates/bevy_ecs/src/lib.rs
#	crates/bevy_ecs/src/schedule/mod.rs
#	crates/bevy_ecs/src/schedule/stage.rs
#	crates/bevy_ecs/src/schedule/stage_executor.rs
#	crates/bevy_ecs/src/schedule/state.rs
#	crates/bevy_ecs/src/system/into_system.rs
#	crates/bevy_ecs/src/system/into_thread_local.rs
#	crates/bevy_ecs/src/system/system.rs
#	crates/bevy_ecs/src/system/system_chaining.rs
#	crates/bevy_ecs/src/system/system_param.rs
#	crates/bevy_input/src/lib.rs
#	crates/bevy_render/src/lib.rs
#	crates/bevy_render/src/render_graph/system.rs
#	crates/bevy_scene/src/lib.rs
#	crates/bevy_transform/src/hierarchy/hierarchy_maintenance_system.rs
#	crates/bevy_transform/src/transform_propagate_system.rs
#	crates/bevy_ui/src/lib.rs
#	crates/bevy_ui/src/update.rs
#	crates/bevy_winit/src/lib.rs
#	examples/2d/texture_atlas.rs
#	examples/ecs/ecs_guide.rs
#	examples/ecs/fixed_timestep.rs
#	examples/ecs/state.rs

+++ crates/bevy_input/src/mouse.rs
@@ -17,7 +17,7 @@ pub enum MouseButton {
     Left,
     Right,
     Middle,
-    Other(u8),
+    Other(u16),
 }
 
 /// A mouse motion event
